Extract information validation utilities to avoid cyclomatic dependency
ContactInfoStep -> AdditionalInformationStep -> BookingInfoStep -> ContactInfoStep
